requirements.txt
# --- For-Bot-Working --------- #
pyrofork==2.3.45
TgCrypto
pyromod==1.5
schedule
apscheduler
# --- For-Database ------------ #
pymongo
motor
dnspython
# --- For-Web-Response ------- #
aiohttp
python-dotenv
# --- For-Helpers ------------ #
pytz
requests
beautifulsoup4
shortzy
aiofiles
humanize
# --- For-Terabox-Download ------- #
uvloop==0.21.0
aria2p
gunicorn
gevent
flask
